requirements.txt
Pillow>=10.0
numpy>=1.26
requests>=2.31
google-genai>=1.0
opencv-python-headless>=4.8
ultralytics>=8.0
